CI fix round 5: fix kpti.h incomplete struct type, pic.c/vga.c narrowing convs
- kpti.h: forward-declare struct process to fix Clang conflicting types
- pic.c: cast uint8_t for outb() arg to fix int→unsigned narrowing
- vga.c: cast uint16_t pos and uint8_t color to fix int→unsigned narrowing

+++ src/drivers/pic.c
@@ -39,5 +39,5 @@ void pic_unmask(uint8_t irq) {
     uint16_t port;
     if (irq < 8) { port = PIC1_DATA; }
     else { port = PIC2_DATA; irq -= 8; }
-    outb(port, inb(port) & ~(1 << irq));
+    outb(port, (uint8_t)(inb(port) & ~(1 << irq)));
 }
